refactor!: rename base-feature to base-presentation
BREAKING CHANGE: Add TYPESAFE_PROJECT_ACCESSORS to enable accessing project modules in build.gradle files with a more robust way

Update readme

Put coroutines dependency group into base dependency group

+++ core/database/src/main/kotlin/core/database/feature/movies/popular/PopularMovieEntity.kt
@@ -10,7 +10,7 @@ class PopularMovieEntity(
     val page: Int,
     val title: String,
     @ColumnInfo(name = "poster_path") val posterPath: String?,
-    @ColumnInfo(name = "vote_average") val voteAverage: Double
+    @ColumnInfo(name = "vote_average") val voteAverage: Float
 ) {
     @PrimaryKey(autoGenerate = true)
     var primaryKey: Int = 0
