+++ feature/movies/data/src/main/kotlin/feature/movies/data/remote/response/PopularMoviesResponse.kt
@@ -16,5 +16,5 @@ data class PopularMovieDto(
     val id: Int,
     val title: String,
     @SerialName("poster_path") val posterPath: String?,
-    @SerialName("vote_average") val voteAverage: Double
+    @SerialName("vote_average") val voteAverage: Float
 )
